--- src/main/java/com/example/oose_23_3_back/managementcontrol/BicycleRepository.java
@@ -1,6 +1,0 @@
-package com.example.oose_23_3_back.managementcontrol;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BicycleRepository extends JpaRepository<Bicycle, Integer> {
-}
