--- src/main/java/com/example/oose_23_3_back/managementcontrol/RentalOfficeRepository.java
@@ -1,6 +1,0 @@
-package com.example.oose_23_3_back.managementcontrol;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RentalOfficeRepository extends JpaRepository<RentalOffice, Integer> {
-}
