--- src/main/java/com/example/oose_23_3_back/managementcontrol/RentalOfficeService.java
@@ -1,14 +1,0 @@
-package com.example.oose_23_3_back.managementcontrol;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class RentalOfficeService {
-    private final RentalOfficeRepository rentalOfficeRepository;
-
-    public void rentalOfficeCreate(RentalOffice rentalOffice) {
-        this.rentalOfficeRepository.save(rentalOffice);
-    }
-}
