+++ src/main/java/com/example/oose_23_3_back/managementcontrol/repository/RentalOfficeRepository.java
@@ -1,0 +1,7 @@
+package com.example.oose_23_3_back.managementcontrol.repository;
+
+import com.example.oose_23_3_back.managementcontrol.entity.RentalOffice;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RentalOfficeRepository extends JpaRepository<RentalOffice, Integer> {
+}
