--- src/main/java/com/example/oose_23_3_back/membercontrol/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.example.oose_23_3_back.membercontrol;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Integer> {
-    Member findByIdentification(String id);
-    Member findByPassword(String password);
-    Member findByRentalPassword(String rentalPassword);
-}
