+++ src/main/java/com/example/oose_23_3_back/membercontrol/control/MemberControl.java
@@ -1,0 +1,18 @@
+package com.example.oose_23_3_back.membercontrol.control;
+
+import com.example.oose_23_3_back.membercontrol.service.MemberService;
+import com.example.oose_23_3_back.membercontrol.entity.Member;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+public class MemberControl {
+    private final MemberService memberService;
+
+    @GetMapping("/memberRead")
+    public Member memberSelect(Integer memberNumber) {
+        return this.memberService.getMember(memberNumber);
+    }
+}
