--- src/main/java/com/example/oose_23_3_back/noticecontrol/PostRepository.java
@@ -1,6 +1,0 @@
-package com.example.oose_23_3_back.noticecontrol;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<Post, Integer> {
-}
