+++ src/main/java/com/example/oose_23_3_back/noticecontrol/service/PostService.java
@@ -1,0 +1,22 @@
+package com.example.oose_23_3_back.noticecontrol.service;
+
+import com.example.oose_23_3_back.noticecontrol.repository.PostRepository;
+import com.example.oose_23_3_back.noticecontrol.entity.Post;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Service
+public class PostService {
+    private final PostRepository postRepository;
+
+    public void postNotice(Post post) {
+        this.postRepository.save(post);
+    }
+
+    public List<Post> readAllPost() {
+        return postRepository.findAll();
+    }
+}
